fix error in changelog and add missing size_t cast
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@1702 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/winio.c
@@ -926,7 +926,7 @@ size_t strnlenpt(const char *buf, size_t size)
 /* How many columns wide is buf? */
 size_t strlenpt(const char *buf)
 {
-    return strnlenpt(buf, -1);
+    return strnlenpt(buf, (size_t)-1);
 }
 
 void blank_bottombars(void)
